\name{paramecium}
\alias{paramecium}
\docType{data}
\title{
Paramecium Example
}
\description{
Paramecium example.
}
\usage{data(paramecium)}
\format{
Formal class \code{\link[dcmle]{dcFit}}.
}
\details{
FIXME
}
\source{
Data and model taken from Ponciano et al. 2009.
Ecology 90, 356-362.
}
\references{
Ponciano et al. 2009. Ecology 90, 356-362.
}
\examples{
\dontrun{
dcexample(paramecium, n.clones=1:2, n.iter=1000)
}
}
\keyword{datasets}
