\name{rats}
\alias{rats}
\docType{data}
\title{
Rats Example
}
\description{
Rats example
}
\usage{data(rats)}
\format{
Formal class \code{\link[dcmle]{dcFit}}.
}
\details{
FIXME
}
\source{
BUGS Manual and references there.
}
\references{
BUGS Manual and references there.
}
\examples{
\dontrun{
dcexample(rats, n.clones=1:2, n.iter=1000)
}
}
\keyword{datasets}
