\name{seeds}
\alias{seeds}
\docType{data}
\title{
Seeds Example
}
\description{
Seeds example.
}
\usage{data(seeds)}
\format{
Formal class \code{\link[dcmle]{dcFit}}.
}
\details{
FIXME
}
\source{
BUGS Manual and references there.
}
\references{
BUGS Manual and references there.
}
\examples{
\dontrun{
dcexample(seeds, n.clones=1:2, n.iter=1000)
}
}
\keyword{datasets}
